refactor: move components to the components directory
