+++ src/components/results-section/results-section.css
